Changed the script
-> First checking if node present or not if not present then install
node
-> Then check if /var/www/demo_react is present or not
-> If present then remove otherwise not do anything

+++ scripts/cleanup.sh
@@ -1,2 +1,20 @@
 #! /bin/bash
-sudo rm -rf /var/www/demo_react
+
+# Check if node is present or not on the system
+
+if ! command -v node &> /dev/null; then
+    echo "Node is not on the system. Installing the latest version of node..."
+    sudo apt-get install -y nodejs
+fi
+
+if command -v node &> /dev/null; then
+    echo "Node is installed."
+
+    # Checking if /var/www/demo_react exists
+    if [-d "/var/www/demo_react" ]; then
+        echo "Removing /var/www/demo_react folder ..."
+        sudo rm -rf /var/www/demo_react
+    else
+        echo "/var/www/demo_react folder not found."
+    fi
+fi
